backend/python/fastapi_tutorial/requirements.txt
anyio==3.7.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==41.0.1
dnspython==2.3.0
ecdsa==0.18.0
email-validator==2.0.0.post2
exceptiongroup==1.1.1
fastapi==0.97.0
h11==0.14.0
httptools==0.5.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
Naked==0.1.32
passlib==1.7.4
pyasn1==0.5.0
pycparser==2.21
pycryptodome==3.18.0
pydantic==1.10.9
PyMySQL==1.0.3
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0
requests==2.31.0
rsa==4.9
shellescape==3.8.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.16
starlette==0.27.0
typing_extensions==4.6.3
urllib3==2.0.3
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
